create listClientInfo

+++ FirstExcelExtract/Models/ClientInfo.cs
@@ -10,6 +10,6 @@ namespace FirstExcelExtract.Models
     {
         public string Name { get; set; }
         public string Gender { get; set; }
-        public int Age { get; set; }
+        public double Age { get; set; }
     }
 }
